fazendo excecoes e tratamento para analise de credito nao encontrada e cliente invalido

+++ src/main/java/com/client/portador/apicreditanalysis/dto/CreditAnalysisDto.java
@@ -1,11 +1,10 @@
 package com.client.portador.apicreditanalysis.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreditAnalysisDto(
         UUID id,
         UUID clientId,
-        BigDecimal approvedLimit
+        Double approvedLimit
 ) {
 }
